+++ source/dynamical_systems/test/tests/test_factory.cpp
@@ -3,8 +3,9 @@
 #include "dynamical_systems/DynamicalSystemFactory.hpp"
 #include "dynamical_systems/exceptions/InvalidParameterException.hpp"
 
-#include "state_representation/parameters/Parameter.hpp"
+#include "state_representation/space/cartesian/CartesianState.hpp"
 #include "state_representation/robot/JointState.hpp"
+#include "state_representation/parameters/Parameter.hpp"
 
 using namespace state_representation;
 
@@ -24,7 +25,6 @@ TEST(DSFactoryTest, CreateDS) {
   auto joint_ds = dynamical_systems::DynamicalSystemFactory<JointState>::create_dynamical_system(
       dynamical_systems::DynamicalSystemFactory<JointState>::DYNAMICAL_SYSTEM::NONE
   );
-  joint_ds->set_base_frame(JointState::Zero("robot", 3));
   ASSERT_NO_THROW(auto res = joint_ds->evaluate(JointState::Random("robot", 3)));
   EXPECT_TRUE(joint_ds->evaluate(JointState::Random("robot", 3)).is_empty());
   EXPECT_EQ(joint_ds->get_parameters().size(), 0);